AgenticAIOrchestrator/requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
plotly==5.17.0
pandas>=2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
psutil==5.9.6
asyncio-mqtt==0.16.1
slack-sdk==3.26.1
sendgrid==6.10.0
blinker
psycopg2-binary
python-decouple

# Optional: External agentic system integrations
# Uncomment these if you want to use CrewAI, LangGraph, etc.
# crewai==0.1.0
# langchain==0.1.0
# langchain-core==0.1.0
# langgraph==0.0.20 